+++ sgn/lib/screens/home_screen/widgets/stories/stories_container.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_mobx/flutter_mobx.dart';
 import 'package:provider/provider.dart';
-import 'package:sgn/screens/home_screen/widgets/story.dart';
+import 'package:sgn/screens/home_screen/widgets/stories/story_thumbnail.dart';
 import 'package:sgn/stores/stories_store.dart';
 
 class StoriesContainer extends StatelessWidget {
