+++ sgn/lib/screens/home_screen/widgets/stories/story_thumbnail.dart
